Move classes related to YoutubeLinksProcessor into separate module "components"
