+++ components/src/main/kotlin/by/jprof/telegram/components/youtube/keyboards.kt
@@ -1,10 +1,10 @@
-package by.jprof.telegram.opinions.webhook.processors
+package by.jprof.telegram.components.youtube
 
-import by.jprof.telegram.opinions.webhook.entity.DOWN_VOTE
-import by.jprof.telegram.opinions.webhook.entity.UP_VOTE
-import by.jprof.telegram.opinions.webhook.entity.Votes
-import by.jprof.telegram.opinions.webhook.entity.downVotes
-import by.jprof.telegram.opinions.webhook.entity.upVotes
+import by.jprof.telegram.components.entity.DOWN_VOTE
+import by.jprof.telegram.components.entity.UP_VOTE
+import by.jprof.telegram.components.entity.Votes
+import by.jprof.telegram.components.entity.downVotes
+import by.jprof.telegram.components.entity.upVotes
 import com.github.insanusmokrassar.TelegramBotAPI.types.buttons.InlineKeyboardButtons.CallbackDataInlineKeyboardButton
 
 fun votingKeyBoard(votes: Votes, votesId: String) = listOf(
