+++ settings.gradle.kts
@@ -2,3 +2,4 @@ rootProject.name = "opinions-bot"
 
 include(":webhook")
 include(":inside-java-podcast-poster")
+include(":components")
