+++ webhook/build.gradle.kts
@@ -21,24 +21,18 @@ val intTestImplementation: Configuration by configurations.getting {
 configurations["intTestRuntimeOnly"].extendsFrom(configurations.runtimeOnly.get())
 
 dependencies {
+    implementation(project(":components"))
     implementation(kotlin("stdlib-jdk8"))
-    implementation(platform("org.apache.logging.log4j:log4j-bom:2.13.1"))
     implementation(platform("com.fasterxml.jackson:jackson-bom:2.10.3"))
-    implementation(platform("software.amazon.awssdk:bom:2.11.9"))
     implementation("com.fasterxml.jackson.core:jackson-databind")
     implementation("com.amazonaws:aws-lambda-java-events:2.2.7")
     implementation("com.amazonaws:aws-lambda-java-core:1.2.0")
-    implementation("org.apache.logging.log4j:log4j-core")
     implementation("com.amazonaws:aws-lambda-java-log4j2:1.1.0")
-    implementation("org.apache.logging.log4j:log4j-slf4j-impl:2.13.3")
     implementation("org.koin:koin-core:2.1.5")
     implementation("org.jetbrains.kotlinx:kotlinx-coroutines-core:1.3.9")
     implementation("org.jetbrains.kotlinx:kotlinx-coroutines-jdk8:1.3.9")
     implementation("org.jetbrains.kotlinx:kotlinx-serialization-core:1.0.0-RC")
-    implementation("software.amazon.awssdk:dynamodb")
     implementation("com.google.api-client:google-api-client:1.23.0")
-    implementation("com.google.apis:google-api-services-youtube:v3-rev222-1.25.0")
-    implementation("com.github.insanusmokrassar:TelegramBotAPI:0.28.0")
     implementation("org.jsoup:jsoup:1.13.1")
 
     testImplementation(platform("org.junit:junit-bom:5.6.0"))
